move passes to individual files
